refactor(extract): move the Haxe extractor to extractors/haxe.py (verbatim)
Follows graphify/extractors/MIGRATION.md: the 637-line span moves
byte-identically, extract.py keeps a facade re-export, extractors/__init__.py
gains the registry entry, and no test changed.

All four helpers the span reaches for outside itself
(_LANGUAGE_BUILTIN_GLOBALS, _file_stem, _make_id, _read_text) were already in
extractors/base.py, so nothing needed promoting to shared.

Also corrects test_haxe_finds_calls, which asserted a lowercased
'buildrequest' against the fixture's 'buildRequest'. Haxe is case-sensitive
and the two neighbouring assertions in the same test already expect case to
be preserved; the test had never passed.

Submitted to Perforce as CL 134094 and CL 134095; this commit brings the git
fork level with the depot.
